requirements.txt
torch>=2.1
numpy>=1.24
matplotlib>=3.7
pyyaml>=6.0
tqdm>=4.65
pytest>=7.4
